+++ lib/karafka/connection/client.rb
@@ -23,7 +23,7 @@ module Karafka
       # @yieldparam [Array<Kafka::FetchedMessage>] kafka fetched messages
       # @note This will yield with raw messages - no preprocessing or reformatting.
       def fetch_loop
-        settings = ApiAdapter.consuming(consumer_group)
+        settings = ApiAdapter.consumption(consumer_group)
 
         if consumer_group.batch_fetching
           kafka_consumer.each_batch(*settings) { |batch| yield(batch.messages) }
@@ -66,10 +66,7 @@ module Karafka
       # @param topic [String] topic that we want to pause
       # @param partition [Integer] number partition that we want to pause
       def pause(topic, partition)
-        settings = ApiAdapter.pausing(consumer_group)
-        timeout = settings[:timeout]
-        raise(Errors::InvalidPauseTimeout, timeout) unless timeout.positive?
-        kafka_consumer.pause(topic, partition, settings)
+        kafka_consumer.pause(*ApiAdapter.pause(topic, partition, consumer_group))
       end
 
       # Marks a given message as consumed and commit the offsets
@@ -110,7 +107,7 @@ module Karafka
       # @note We don't cache it internally because we cache kafka_consumer that uses kafka
       #   object instance
       def kafka
-        Kafka.new(*ApiAdapter.client(consumer_group))
+        Kafka.new(*ApiAdapter.client)
       end
     end
   end