+++ lib/karafka/errors.rb
@@ -41,9 +41,6 @@ module Karafka
     # This should never happen and if it does, please contact us
     MissingClient = Class.new(BaseError)
 
-    # Raised when we attemp to pause a partition but the pause timeout is equal to 0
-    InvalidPauseTimeout = Class.new(BaseError)
-
     # Raised when want to hook up to an event that is not registered and supported
     UnregisteredMonitorEvent = Class.new(BaseError)